+++ src/components/KanbanBoard/IssueTitle.tsx
@@ -8,7 +8,7 @@ import { useIssue } from '@/hooks/useIssue';
 import { IssueStateType } from '@/types/issue';
 
 import DeleteHoverBtn from '../Shared/Button/DeleteHoverBtn';
-import DeleteConfirmToast from '../Toast/DeleteConfirmToast';
+import DeleteConfirmToast from '../Shared/Toast/DeleteConfirmToast';
 import IssueInput from './IssueInput';
 
 type IssueTitleProps = {
